refactor: extract FFmpegOutputParser and FFmpegProcess from FFmpegWrapper (PR 1)
